requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-ranged-response==0.2.0
django-recaptcha==4.1.0
django-simple-captcha==0.6.2
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
whitenoise==6.9.0
